Refactor image handling in Dish and Drink components; update services and controllers for image URL management
- Updated DishCard and DrinkCard components to use imageUrl and handleImageChange directly.
- Refactored CardImage and CardImageSection components to streamline image rendering and editing.
- Replaced IngredientCreationModal with DrinkCreationModal for drink item creation.
- Modified services (dishesService, drinksService, ingredientsService) to handle image URLs instead of base64 strings.
- Updated database models for Dish and Drink to use imageUrl instead of image.
- Adjusted backend image upload logic to accept image name and return image URL.
- Removed unnecessary image handling functions from ingredientsController.

+++ client/src/app/components/generic/card/Card.tsx
@@ -23,6 +23,7 @@ interface CardProps<T extends BaseItem> {
   keyField: string,
   canDelete?: boolean,
   canSave?: (newItem: T) => Promise<boolean>;
+  deleteItem: (key: string) => void;
   service: CrudService<T>,
   cardComponent: React.ComponentType<CardComponentProps<T>>;
 };
@@ -32,6 +33,7 @@ export default function Card<T extends BaseItem>({
   keyField,
   canDelete=true,
   canSave,
+  deleteItem,
   service,
   cardComponent: CardComponent,
 }: CardProps<T>) {
@@ -110,22 +112,10 @@ export default function Card<T extends BaseItem>({
       confirmButtonText: "DELETE",
       cancelButtonText: "Undo"
     }).then((result) => {
-      if (result.isConfirmed) {
+      if (result.isConfirmed && componentKey !== "") {
         deleteItem(componentKey);
       }
     });
-  }
-
-  // DELETES ITEM
-  async function deleteItem(componentKey: string) {
-    if (componentKey !== "") {
-      try {
-        await service.deleteItem(componentKey);
-        await service.fetchItems();
-      } catch (error) {
-        console.error(error);
-      }
-    }
   }
 
   function handleCheckChange(event: React.ChangeEvent<HTMLInputElement>, fieldName: string) {
